requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
psycopg2-binary==2.9.3
spacy==3.2.1
openai==0.10.2
sentence-transformers==2.2.0
pandas==1.3.4
requests==2.26.0
beautifulsoup4==4.10.0